requirements.txt
awsglue-dev==2021.12.30
boto3==1.34.122
botocore==1.34.122
configparser==7.0.0
fsspec==2024.6.0
func-timeout==4.3.5
jmespath==1.0.1
mysql-connector-python==8.4.0
numpy==1.26.4
pandas==2.2.2
py4j==0.10.9.7
PyAthena==3.8.3
pyspark==3.5.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
s3transfer==0.10.1
six==1.16.0
SQLAlchemy==2.0.30
tenacity==8.3.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.18
